requirements.txt
adbutils==0.8.3
apkutils2==1.0.0
atomicwrites==1.4.0
attrdict==2.0.1
attrs==20.3.0
cached-property==1.5.2
certifi==2020.12.5
cffi==1.14.4
chardet==3.0.4
cigam==0.0.3
click==7.1.2
colorama==0.4.4
coverage==5.4
crypto==1.4.1
cryptography==3.3.1
Cython==0.29.21
ddt==1.4.1
decorator==4.4.2
Deprecated==1.2.11
deprecation==2.1.0
dubbo-telnet==1.0.1b3
et-xmlfile==1.0.1
facebook-wda==1.3.2
fast-json==0.3.2
Flask==1.1.2
humanize==3.2.0
idna==2.10
iniconfig==1.1.1
itsdangerous==1.1.0
jdcal==1.4.1
Jinja2==2.11.2
jpype1==1.2.0
kazoo==2.8.0
logzero==1.6.3
lxml==4.6.2
MarkupSafe==1.1.1
mysql-connector-python==8.0.22
Naked==0.1.31
numpy==1.19.5
opencv-python==4.5.1.48
openpyxl==3.0.6
packaging==20.8
pandas==1.2.2
Pillow==8.1.0
pluggy==0.13.1
progress==1.5
protobuf==3.14.0
py==1.10.0
pyasn1==0.4.8
pycparser==2.20
pycryptodome==3.9.9
pyelftools==0.27
pyjnius==1.3.0
PyMySQL==0.10.1
pyparsing==2.4.7
pytest==6.2.2
pytest-cov==2.11.1
pytest-html==3.1.1
pytest-instafail==0.4.2
pytest-metadata==1.11.0
pytest-parallel==0.1.0
pytest-rerunfailures==9.1.1
pytest-timeout==1.4.2
python-dateutil==2.8.1
pytils==0.3
pytz==2021.1
pywin32==300
PyYAML==5.4.1
redis==3.5.3
redis-py-cluster==2.1.0
requests==2.25.0
retry==0.9.2
#robotframework @ file:///D:/software%20backup/Python%E6%89%A9%E5%B1%95%E5%8C%85/robotframework-3.2.2-py2.py3-none-any.whl
robotframework-ride==1.7.4.2
rsa==4.7
selenium==3.141.0
shellescape==3.8.1
simplejson==3.17.2
six==1.15.0
tblib==1.7.0
toml==0.10.2
tools==0.1.9
tornado==6.1
uiautomator2==2.12.1
ujson==4.0.2
urllib3==1.26.2
weditor==0.6.3
Werkzeug==1.0.1
whichcraft==0.6.1
wrapt==1.12.1
xmltodict==0.12.0
